+++ pyPRISM/__init__.py
@@ -60,7 +60,12 @@ from pyPRISM import util
 
 from pyPRISM.version import *
 
-
-
-
+def test():
+    r"""
+    Run all tests using pytest.
+    """
+    import os
+    import pytest
+    path = os.path.split(__file__)[0]
+    pytest.main(['-x',path])