+++ CHANGES.md
@@ -1,10 +1,11 @@
 Change Log
 ==========
 
-v1.0.0
+v1.0.2
 ------
-- Date: 18/04/01
-- Initial Release! 
+- Date:  18/04/04
+- Moved tests back under pyPRISM
+- Added test() method to pyPRISM/__init__.py
 
 v1.0.1
 ------
@@ -12,3 +13,9 @@ v1.0.1
 - Added LICENSE to MANIFEST.in
     - Yes, this required a version bump
     - See here: https://github.com/pypa/packaging-problems/issues/74
+
+v1.0.0
+------
+- Date: 18/04/01
+- Initial Release! 
+